Rename spirv_fn_generator to spirv_atomic_fn_declarations.
